refavtor(vision main): Kept the script that manages the servos on the teensy rather than the one that used the GPIO pins on the Rasp
