+++ visao_rasp/Main Visao/raspi_visao.py
@@ -2,24 +2,24 @@
 """
 raspi_visao.py — Headless bean sorter (dispatcher mode)
 =========================================================
-Camera  : /dev/video_zed  (V4L2, ZED 1344x376)
-Config  : raspi_config.json (full-res coords, auto-scaled 2x at runtime)
-Output  : VISION:FF:L:R lines for dispatcher
+Camera  : /dev/video{CAM_PORT}  (V4L2)
+Config  : raspi_config.json
+Serial  : NONE — prints VISION:FF:L:R lines for dispatcher to forward
+Output  : FPS + hit counts to terminal, VISION tags to stdout
 Ctrl+C  : stop
 """
 
 import cv2
 import numpy as np
 import sys, json, os, time
-from ServoSystem import ServoSystem
 
 # ══════════════════════════════════════════════════════════════════════
 #  CONSTANTS
 # ══════════════════════════════════════════════════════════════════════
 CONFIG_FILE = "raspi_config.json"
+CAM_PORT    = 0
 FRAME_W     = 1344
 FRAME_H     = 376
-DS          = 2
 
 # ══════════════════════════════════════════════════════════════════════
 #  CONFIG
@@ -29,8 +29,6 @@ def load_cfg() -> dict:
         sys.exit(f"[ERROR] {CONFIG_FILE} not found — run debug script first and press S.")
     with open(CONFIG_FILE) as f:
         raw = json.load(f)
-
-    s = DS
     return dict(
         s_min      = raw['s_min'],
         v_dark     = raw['v_dark'],
@@ -38,25 +36,45 @@ def load_cfg() -> dict:
         bg_h_hi    = raw['bg_h_hi'],
         bg_s_min   = raw['bg_s_min'],
         bg_v_min   = raw['bg_v_min'],
-        area_min   = (raw['area_min'] * 10) // (s * s),
-        area_max   = (raw['area_max'] * 10) // (s * s),
-        rad_min    = max(1, raw['rad_min'] // s),
-        rad_max    = max(1, raw['rad_max'] // s),
+        area_min   = raw['area_min'] * 10,
+        area_max   = raw['area_max'] * 10,
+        rad_min    = raw['rad_min'],
+        rad_max    = raw['rad_max'],
         circ_min   = raw['circ_min'] * 0.01,
         circ_max   = raw['circ_max'] * 0.01,
         morph_k    = max(1, raw['morph_k'] | 1),
         morph_iter = max(1, raw['morph_iter']),
-        det_cy_l   = raw['det_cy_l'] // s,
-        det_cy_r   = raw['det_cy_r'] // s,
-        det_thick  = max(1, raw['det_thick'] // s),
-        trigger_x  = raw['trigger_x'] // s,
-        trig_y2    = raw['trig_y2'] // s,
+        det_cy_l   = raw['det_cy_l'],
+        det_cy_r   = raw['det_cy_r'],
+        det_thick  = max(1, raw['det_thick']),
+        trigger_x  = raw['trigger_x'],
+        trig_y2    = raw['trig_y2'],
+        gh_lo      = raw['gh_lo'],
+        gh_hi      = raw['gh_hi'],
+        gs_min     = raw['gs_min'],
+        gv_min     = raw['gv_min'],
+        gfrac      = raw['gfrac'] / 100.0,
     )
+
+# ══════════════════════════════════════════════════════════════════════
+#  KNN CLASSIFIER
+# ══════════════════════════════════════════════════════════════════════
+KNN_SAMPLES = np.array([
+    [ 20, 200, 220], [ 25, 180, 210], [ 15, 220, 230],
+    [ 10, 220, 210], [ 12, 200, 200], [  8, 210, 195],
+    [  5, 200, 180], [  3, 210, 190], [  2, 190, 170],
+    [175, 200, 180], [172, 195, 185], [170, 185, 175],
+    [110, 150,  80], [100, 140,  70], [120, 160,  90], [105, 130,  75],
+    [  0,  20,  30], [  0,  10,  20], [ 15,  15,  25], [  0,   5,  15],
+], dtype=np.float32)
+KNN_LABELS = np.array([0]*12 + [1]*8, dtype=np.int32)
+_knn = cv2.ml.KNearest_create()
+_knn.train(KNN_SAMPLES, cv2.ml.ROW_SAMPLE, KNN_LABELS)
 
 # ══════════════════════════════════════════════════════════════════════
 #  KALMAN GHOST TRACKER
 # ══════════════════════════════════════════════════════════════════════
-GHOST_FRAMES = 0
+GHOST_FRAMES = 10
 
 def _make_kalman():
     kf = cv2.KalmanFilter(4, 2)
@@ -75,7 +93,7 @@ _ghost = {
 
 def update_ghost(side, contour_data):
     g    = _ghost[side]
-    best = next(((cx,cy,ri) for cx,cy,ri,passed in contour_data if passed), None)
+    best = next(((cx,cy,ri) for cx,cy,ri,*_,passed,_ in contour_data if passed), None)
     if best:
         cx, cy, ri = best
         g['kf'].correct(np.array([[np.float32(cx)],[np.float32(cy)]]))
@@ -84,25 +102,25 @@ def update_ghost(side, contour_data):
     else:
         g['ttl'] = max(0, g['ttl'] - 1)
     pred     = g['kf'].predict()
-    g['pos'] = (int(pred[0].item()), int(pred[1].item()))
+    g['pos'] = (int(pred[0]), int(pred[1]))
 
 def check_trigger(contour_data, cfg, side):
     y_lo = min(cfg['trigger_x'], cfg['trig_y2'])
     y_hi = max(cfg['trigger_x'], cfg['trig_y2'])
-    for cx, cy, ri, passed in contour_data:
-        if passed and (cy + ri >= y_lo) and (cy - ri <= y_hi):
+    for cx, cy, ri, cnt, area, passed, circ in contour_data:
+        if passed and (cy+ri >= y_lo) and (cy-ri <= y_hi):
             return True
     g = _ghost[side]
     if g['ttl'] > 0:
         gcy, gri = g['pos'][1], g['r']
-        if (gcy + gri >= y_lo) and (gcy - gri <= y_hi):
+        if (gcy+gri >= y_lo) and (gcy-gri <= y_hi):
             return True
     return False
 
 # ══════════════════════════════════════════════════════════════════════
 #  VISION PIPELINE
 # ══════════════════════════════════════════════════════════════════════
-def build_mask(half, cfg):
+def build_masks(half, cfg):
     hsv      = cv2.cvtColor(half, cv2.COLOR_BGR2HSV)
     raw_cand = ((hsv[:,:,1] >= cfg['s_min']) | (hsv[:,:,2] < cfg['v_dark'])).astype(np.uint8) * 255
     bg_lo    = np.array([cfg['bg_h_lo'], cfg['bg_s_min'], cfg['bg_v_min']], dtype=np.uint8)
@@ -111,26 +129,27 @@ def build_mask(half, cfg):
     k        = cv2.getStructuringElement(cv2.MORPH_ELLIPSE, (cfg['morph_k'], cfg['morph_k']))
     morph    = cv2.morphologyEx(clean, cv2.MORPH_OPEN,  k, iterations=cfg['morph_iter'])
     morph    = cv2.morphologyEx(morph, cv2.MORPH_CLOSE, k, iterations=cfg['morph_iter'])
-    return morph
+    bean_lo  = np.array([cfg['gh_lo'], cfg['gs_min'], cfg['gv_min']], dtype=np.uint8)
+    bean_hi  = np.array([cfg['gh_hi'], 255,           255],           dtype=np.uint8)
+    return dict(hsv=hsv, cand_morph=morph, bean_green=cv2.inRange(hsv, bean_lo, bean_hi))
 
-def detect(mask, cfg, side):
+def detect(half, masks, cfg, side):
     det_cy = cfg['det_cy_l'] if side == 'left' else cfg['det_cy_r']
     det_th = cfg['det_thick']
-    cnts, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    cnts, _ = cv2.findContours(masks['cand_morph'], cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
     out = []
     for cnt in cnts:
         area  = cv2.contourArea(cnt)
         perim = cv2.arcLength(cnt, True)
-        if perim == 0:
-            continue
+        if perim == 0: continue
         circ      = 4 * np.pi * area / (perim ** 2)
-        (x, y), r = cv2.minEnclosingCircle(cnt)
-        cx, cy, ri = int(x), int(y), int(r)
-        passed = (abs(cx - det_cy) <= det_th and
-                  cfg['rad_min']  <= ri   <= cfg['rad_max'] and
-                  cfg['area_min'] <= area <= cfg['area_max'] and
-                  cfg['circ_min'] <= circ <= cfg['circ_max'])
-        out.append((cx, cy, ri, passed))
+        (x,y), r  = cv2.minEnclosingCircle(cnt)
+        cx,cy,ri  = int(x), int(y), int(r)
+        passed    = (abs(cx - det_cy) <= det_th and
+                     cfg['rad_min']  <= ri   <= cfg['rad_max'] and
+                     cfg['area_min'] <= area <= cfg['area_max'] and
+                     cfg['circ_min'] <= circ <= cfg['circ_max'])
+        out.append((cx, cy, ri, cnt, area, passed, circ))
     return out
 
 # ══════════════════════════════════════════════════════════════════════
@@ -138,28 +157,20 @@ def detect(mask, cfg, side):
 # ══════════════════════════════════════════════════════════════════════
 def main():
     cfg = load_cfg()
-    print(f"[Config] loaded {CONFIG_FILE} (DS={DS}x)", file=sys.stderr)
+    print(f"[Config] loaded {CONFIG_FILE}")
 
-    cap = cv2.VideoCapture("/dev/video_zed", cv2.CAP_V4L2)
+    cap = cv2.VideoCapture(CAM_PORT, cv2.CAP_V4L2)
     cap.set(cv2.CAP_PROP_FRAME_WIDTH,  FRAME_W)
     cap.set(cv2.CAP_PROP_FRAME_HEIGHT, FRAME_H)
-    cap.set(cv2.CAP_PROP_BUFFERSIZE, 1)
     if not cap.isOpened():
-        sys.exit("[ERROR] Cannot open /dev/video_zed")
+        sys.exit(f"[ERROR] Cannot open /dev/video{CAM_PORT}")
 
     actual_w = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
     actual_h = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
     half_w   = actual_w // 2
-    ds_w     = half_w // DS
-    ds_h     = actual_h // DS
-    print(f"[Camera] /dev/video_zed  {actual_w}x{actual_h}  half={half_w}x{actual_h}  ds={ds_w}x{ds_h}", file=sys.stderr)
-    print("[Running] Ctrl+C to stop", file=sys.stderr)
-
-    servos = ServoSystem(upper_pin=12, lower_pin=13)
-    servos.intake_home()
-    time.sleep(0.5)          # let servos reach home before starting loop
-    servos.update()
-    print("[Servos] initialized, homed", file=sys.stderr)
+    print(f"[Camera] /dev/video{CAM_PORT}  {actual_w}x{actual_h}  half={half_w}x{actual_h}")
+    print("[Mode] Dispatcher (VISION tags via stdout)")
+    print("[Running] Ctrl+C to stop")
 
     frame_count = 0
     hit_counts  = {'left': 0, 'right': 0}
@@ -168,70 +179,37 @@ def main():
 
     try:
         while True:
-            t_grab = time.monotonic()
             ret, frame = cap.read()
             if not ret or frame is None:
                 continue
-            t_read = time.monotonic()
 
-            halves = [
-                ('left',  cv2.resize(frame[:, :half_w], (ds_w, ds_h), interpolation=cv2.INTER_NEAREST)),
-                ('right', cv2.resize(frame[:, half_w:], (ds_w, ds_h), interpolation=cv2.INTER_NEAREST)),
-            ]
-            t_resize = time.monotonic()
-
-            # ── detect both sides ──
-            hits = {}
+            halves = [('left', frame[:, :half_w]), ('right', frame[:, half_w:])]
+            hits   = {}
             for side, half in halves:
-                mask   = build_mask(half, cfg)
-                cdata  = detect(mask, cfg, side)
+                masks       = build_masks(half, cfg)
+                cdata       = detect(half, masks, cfg, side)
                 update_ghost(side, cdata)
-                hits[side] = check_trigger(cdata, cfg, side)
+                hits[side]  = check_trigger(cdata, cfg, side)
                 if hits[side]:
                     hit_counts[side] += 1
 
-            # ── command servos individually ──
-            if hits['left']:
-                servos.intake_upper_deploy()
-            else:
-                servos.intake_upper_home()
+            # Output vision data for dispatcher to forward
+            # Format: VISION:FF:LL:RR (hex bytes)
+            print(f"VISION:FF:{int(hits['left']):02X}:{int(hits['right']):02X}", flush=True)
 
-            if hits['right']:
-                servos.intake_lower_deploy()
-            else:
-                servos.intake_lower_home()
-
-            servos.update()       # ← detach timer fires here
-
-            t_proc = time.monotonic()
-
-            # ── dispatcher output (uncomment when ready) ──
-            # bitfield = (int(hits['right']) & 1) | ((int(hits['left']) & 1) << 1)
-            # print(f"VISION:{bitfield:02X}", flush=True)
-
-            # ── FPS logging ──
             frame_count += 1
             now = time.time()
             if now - t_last >= 1.0:
                 fps = frame_count / (now - t0)
                 l_str = "L" if hits['left'] else " "
                 r_str = "R" if hits['right'] else " "
-                read_ms   = (t_read   - t_grab)   * 1000
-                resize_ms = (t_resize - t_read)    * 1000
-                proc_ms   = (t_proc   - t_resize)  * 1000
-                total_ms  = (t_proc   - t_grab)    * 1000
-                print(f"[FPS] {fps:5.1f}  |  {l_str} {r_str}  |  "
-                      f"L:{hit_counts['left']:4d}  R:{hit_counts['right']:4d}  |  "
-                      f"read:{read_ms:.0f} resize:{resize_ms:.0f} proc:{proc_ms:.0f} "
-                      f"total:{total_ms:.0f}ms", file=sys.stderr)
+                print(f"[FPS] {fps:5.1f}  |  {l_str} {r_str}  |  L hits: {hit_counts['left']:4d}  R hits: {hit_counts['right']:4d}")
                 t_last = now
 
     except KeyboardInterrupt:
         elapsed = time.time() - t0
-        print(f"\n[Stopped]  {frame_count} frames in {elapsed:.1f}s  "
-              f"avg {frame_count/elapsed:.1f} fps", file=sys.stderr)
+        print(f"\n[Stopped]  {frame_count} frames in {elapsed:.1f}s  avg {frame_count/elapsed:.1f} fps")
     finally:
-        servos.close()
         cap.release()
 
 if __name__ == "__main__":
